+++ R/methods_rstan.R
@@ -60,7 +60,7 @@ model_parameters.stanfit <- function(model,
 
   attr(params, "parameter_info") <- insight::clean_parameters(model)
   attr(params, "object_name") <- insight::safe_deparse_symbol(substitute(model))
-  class(params) <- c("parameters_stan", "parameters_model", "see_parameters_model", class(params))
+  class(params) <- c("parameters_model", "see_parameters_model", class(params))
 
   params
 }