+++ tests/testthat/test-GLMMadaptive.R
@@ -135,7 +135,7 @@ if (
     {
       si <- Sys.info()
       if (!is.null(si["sysname"])) {
-        si["sysname"] == "Windows" || grepl("^mingw", R.version$os)
+        si["sysname"] == "Windows" || startsWith(R.version$os, "mingw")
       } else {
         FALSE
       }